Change measurement seconds to float
We have to deal with fractional seconds in the nadir case, so we change
the variable to float to not to lose precision.  In the limb case, they
will automagically converted to int.  However, for the retrieval this will
not matter much, if at all.

+++ Messung.h
@@ -82,7 +82,7 @@ public:
 	int m_Tag;
 	int m_Stunde;
 	int m_Minute;
-	int m_Sekunde;
+	float m_Sekunde;
 	// Geolocation
 	double m_Latitude_Sat;   // hierfür muss das Programm noch geändert werden
 	double m_Longitude_Sat;
